Year 2038 fixes

+++ Core/Resources/Db/MySQL/db.sql
@@ -13,7 +13,7 @@ CREATE TABLE cards (
     addressbookid INT(11) UNSIGNED NOT NULL,
     carddata MEDIUMBLOB,
     uri VARBINARY(200),
-    lastmodified INT(11) UNSIGNED,
+    lastmodified BIGINT(11) UNSIGNED,
     etag VARBINARY(32),
     size INT(11) UNSIGNED NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
@@ -32,12 +32,12 @@ CREATE TABLE calendarobjects (
     calendardata MEDIUMBLOB,
     uri VARBINARY(200),
     calendarid INTEGER UNSIGNED NOT NULL,
-    lastmodified INT(11) UNSIGNED,
+    lastmodified BIGINT(11) UNSIGNED,
     etag VARBINARY(32),
     size INT(11) UNSIGNED NOT NULL,
     componenttype VARBINARY(8),
-    firstoccurence INT(11) UNSIGNED,
-    lastoccurence INT(11) UNSIGNED,
+    firstoccurence BIGINT(11) UNSIGNED,
+    lastoccurence BIGINT(11) UNSIGNED,
     uid VARBINARY(200),
     UNIQUE(calendarid, uri)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
@@ -89,7 +89,7 @@ CREATE TABLE calendarsubscriptions (
     striptodos TINYINT(1) NULL,
     stripalarms TINYINT(1) NULL,
     stripattachments TINYINT(1) NULL,
-    lastmodified INT(11) UNSIGNED,
+    lastmodified BIGINT(11) UNSIGNED,
     UNIQUE(principaluri, uri)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
@@ -98,7 +98,7 @@ CREATE TABLE schedulingobjects (
     principaluri VARBINARY(255),
     calendardata MEDIUMBLOB,
     uri VARBINARY(200),
-    lastmodified INT(11) UNSIGNED,
+    lastmodified BIGINT(11) UNSIGNED,
     etag VARBINARY(32),
     size INT(11) UNSIGNED NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
